requirements.txt
Flask==3.1.0
loguru==0.7.3
python-osc==1.9.3
PyYAML==6.0.2
pyserial==3.5
click==8.1.7